requirements.txt
requests
pandas
schedule
python-dotenv
supabase
plotly
streamlit
gTTS
numpy
